welcome_to/requirements.txt
certifi==2018.4.16
chardet==3.0.4
google-api-python-client==1.6.6
httplib2==0.11.3
idna==2.6
oauth2client==4.1.2
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytesseract==0.2.0
requests==2.18.4
rsa==3.4.2
selenium==3.11.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
